CV-Modul: SDRAM funktioniert @200MHz, WS2812 LEDs lassen sich per DMA einschalten, DCach cleanup benötigt

+++ CV_Modul_F730ZB/Core/Inc/globals.h
@@ -32,7 +32,7 @@ extern uint8_t pic_captured;
 extern uint8_t pic_written;
 extern uint8_t led_val;
 extern uint8_t led_val_changed;
-extern uint16_t cursorLine;
+extern uint32_t cursorLine;
 extern uint16_t actual_menu;
 // Framebuffer for capuring images and writing the display
